backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
fastapi-users[sqlalchemy]==12.1.2
python-dotenv==1.0.1
sqlalchemy==2.0.27
sqlmodel==0.0.14
fastapi-mail==1.4.1
resend==0.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosqlite==0.19.0
pydantic[email]==2.6.1
pydantic-settings==2.1.0
email-validator<2.1,>=1.1.0
watchfiles==0.21.0
requests==2.31.0
greenlet==3.0.3
typing-extensions>=4.8.0
